refactor(workspace): answer the two-axis review of the normalization work
A standards pass and a spec pass over the branch, applied. Most of it is naming
and placement, but one finding turned out to be a false claim in a comment, and
that one is pinned rather than reworded.

`leading_zero` said its unit-less branch was unreachable. It is not.
`parse_js_float` skips leading JavaScript whitespace and `unit` does not, and
the value scanner does not treat U+00A0 as a space, so a word can begin with
that character, read back as a number, and split into no dimension at all --
whereupon the re-spelling drops the character and the unit with it, and
`\u{a0}0.5px` comes back as `.5`. That is the reference compiler's ternary doing
what it does there, so it is now a test rather than a sentence, and it is the
reason this normalizer keeps a walk of its own.

The other three that open the same way do not all fold together either.
`zero_dimensions` tracks function boundaries across every node, words included,
so it has to see what a word filter would remove. The two that are left --
`timings` and `font_size_px_to_rem` -- now share `walk_dimensions`, and the
exclusions are recorded beside it so the next reader does not have to rediscover
them the same way.

The parity harness asserted parsed JSON into shape and read `stylex` metadata
through a cast. Corpus files now narrow at the read and name the file that was
wrong, instead of failing somewhere inside the loop that consumes them.
`harvest.ts` was doing three jobs; reading Rust sources moved to `rust-source.ts`
and entry identity to `declaration.ts`, so adding a literal shape touches the
extractors alone. The harvest is unchanged by it: same 734 declarations, same
ids, only the line numbers this commit's new test shifted.

The rest is prose. The nine ports count their position against upstream's
sequence and never did count against the local one, which runs a tenth pass
upstream has no equivalent for; every header now says so. The exports with no
caller stay, and `lib.rs` says why -- a missing one is a difference from the
thing being ported. The licence paragraph in the README was pasted twice.
