+++ crates/stylex-css/src/css/normalizers/timings.rs
@@ -1,7 +1,11 @@
-//! Ported normalizer 4 of 9. See `normalize_value.rs` for the ordered list.
+//! Ported normalizer 4 of 9 in upstream's sequence, which is what that count
+//! names. See `normalize_value.rs` for the order the passes run in here — a
+//! tenth pass that upstream has no equivalent for runs among them.
 
-use postcss_value_parser::{NodeKind, ValueParser, unit};
+use postcss_value_parser::ValueParser;
 use stylex_utils::number::{parse_js_float, to_js_string};
+
+use super::dimensions::walk_dimensions;
 
 /// Rewrites a duration of ten milliseconds or more as seconds.
 ///
@@ -12,30 +16,16 @@ use stylex_utils::number::{parse_js_float, to_js_string};
 /// Anything under ten milliseconds is left alone, so `9ms` stays `9ms` — the
 /// conversion would spell it `.009s`, which is longer than what it replaced.
 pub fn normalize_timings(ast: &mut ValueParser, _key: &str) {
-  ast.walk(
-    |node, _| {
-      if node.kind != NodeKind::Word {
-        return true;
-      }
+  walk_dimensions(ast, |node, dimension| {
+    let value = match parse_js_float(&node.value) {
+      Some(value) => value,
+      None => return,
+    };
 
-      let value = match parse_js_float(&node.value) {
-        Some(value) => value,
-        None => return true,
-      };
+    if dimension.unit != "ms" || value < 10.0 {
+      return;
+    }
 
-      let dimension = match unit(&node.value) {
-        Some(dimension) => dimension,
-        None => return true,
-      };
-
-      if dimension.unit != "ms" || value < 10.0 {
-        return true;
-      }
-
-      node.value = format!("{}s", to_js_string(value / 1000.0));
-
-      true
-    },
-    false,
-  );
+    node.value = format!("{}s", to_js_string(value / 1000.0));
+  });
 }
