+++ crates/stylex-rs-compiler/parity/lib/corpus.ts
@@ -10,8 +10,9 @@
 import fs from 'node:fs';
 import path from 'node:path';
 
-import { declarationKey } from './harvest.js';
-import type { CorpusFile, LoadedCorpusEntry } from './types.js';
+import { declarationKey } from './declaration.js';
+import { arrayAt, stringAt } from './guards.js';
+import type { CorpusEntry, CorpusFile, LoadedCorpusEntry } from './types.js';
 
 /** Load order is the report order, so the reported cases read first. */
 export const CORPUS_FILES = ['reported.json', 'edge.json', 'harvested.json'] as const;
@@ -30,7 +31,7 @@ export function loadCorpus(corpusDir: string): LoadedCorpusEntry[] {
       );
     }
 
-    const file = JSON.parse(fs.readFileSync(filePath, 'utf8')) as CorpusFile;
+    const file = corpusFileFrom(JSON.parse(fs.readFileSync(filePath, 'utf8')), filePath);
     for (const entry of file.entries) {
       // A hand-written case that also appears in the harvest keeps its own
       // note and origin; the harvested duplicate adds nothing.
@@ -42,4 +43,35 @@ export function loadCorpus(corpusDir: string): LoadedCorpusEntry[] {
   }
 
   return entries;
+}
+
+/** Narrow parsed JSON to a corpus file, naming the file when it is not one. */
+function corpusFileFrom(raw: unknown, filePath: string): CorpusFile {
+  const set = stringAt(raw, 'set');
+  const description = stringAt(raw, 'description');
+  const entries = arrayAt(raw, 'entries');
+  if (set === undefined || description === undefined || entries === undefined) {
+    throw new Error(`Corpus file malformed: ${filePath} — expected { set, description, entries }.`);
+  }
+
+  return {
+    set,
+    description,
+    entries: entries.map((entry, index) => corpusEntryFrom(entry, filePath, index)),
+  };
+}
+
+function corpusEntryFrom(raw: unknown, filePath: string, index: number): CorpusEntry {
+  const id = stringAt(raw, 'id');
+  const property = stringAt(raw, 'property');
+  const value = stringAt(raw, 'value');
+  const origin = stringAt(raw, 'origin');
+  if (id === undefined || property === undefined || value === undefined || origin === undefined) {
+    throw new Error(
+      `Corpus entry ${index} malformed in ${filePath} — expected { id, property, value, origin }.`
+    );
+  }
+
+  const note = stringAt(raw, 'note');
+  return { id, property, value, origin, ...(note === undefined ? {} : { note }) };
 }
